scripts/requirements.txt
Pillow
psycopg2-binary